+++ src/components/manage/Add/Add.jsx
@@ -334,6 +334,7 @@ class Add extends Component {
           {this.state.isClient && (
             <Portal node={document.getElementById('toolbar')}>
               <Toolbar
+                activity="add"
                 pathname={this.props.pathname}
                 hideDefaultViewButtons
                 inner={